Fix negative duration in Activity
Fix overflow due to cast from int to short while extending the maximum
duration to the one in PatternActivity.

+++ mobitopp/src/main/java/edu/kit/ifv/mobitopp/simulation/activityschedule/Activity.java
@@ -64,8 +64,8 @@ public class Activity implements ActivityIfc {
   		byte activityNrOfWeek,
   		ActivityType activityType,
   		Time startDate,
-  		short duration,
-  		short observedTripDuration
+  		int duration,
+  		int observedTripDuration
   	) {
   	this(oid, activityNrOfWeek, activityType, startDate, duration, observedTripDuration, 1.0f, 0.0f, 1.0f);
   }
